docker/ari_server/requirements.txt
requests==2.31.0
websocket-client==1.7.0
numpy==1.26.4
soundfile==0.12.1
scipy==1.11.4